clean up session and session launcher cards
