+++ client/src/features/sessionsV2/SessionList/Session.module.scss
